feat(onboarding): implementation of Actions in the app
- Added all the Actions that the user can have in the app
- Organized all screens on their individual package libraries

Technical notes:
- Implementing of Root composable for easier way to test the app and preview
- Data flow and state management to be implemented next
- Ready for backend integration and user preference storage

Part of feature: BALESTATE-7
