+++ onboarding/presentation/src/main/java/com/zanoapps/onboarding/presentation/buyer/amenities/AmenitiesBuyerScreen.kt
@@ -1,4 +1,5 @@
-package com.zanoapps.onboarding.presentation.seller
+package com.zanoapps.onboarding.presentation.buyer.amenities
+
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -13,79 +14,84 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.zanoapps.core.presentation.designsystem.BalkanEstateTheme
 import com.zanoapps.core.presentation.designsystem.Poppins
-import com.zanoapps.core.presentation.designsystem.R
 import com.zanoapps.core.presentation.designsystem.components.BalkanEstateActionButton
 import com.zanoapps.core.presentation.designsystem.components.BalkanEstateOutlinedActionButton
 import com.zanoapps.core.presentation.designsystem.components.GradientBackground
-import com.zanoapps.onboarding.domain.enums.seller.MainGoal
-import com.zanoapps.onboarding.domain.enums.seller.PropertyTypeSeller
-import com.zanoapps.onboarding.domain.enums.seller.SellingTime
+import com.zanoapps.onboarding.domain.enums.buyer.Amenity
 import com.zanoapps.onboarding.presentation.components.BalkanEstateSelectionCard
 import com.zanoapps.onboarding.presentation.components.ProgressBar
 import com.zanoapps.onboarding.presentation.components.SelectionType
 import com.zanoapps.onboarding.presentation.components.SkipSurvey
 
 @Composable
-fun MainGoalScreen(
-    selectedType: MainGoal,
-    onToggleSelection: (MainGoal) -> Unit,
+fun AmenitiesBuyerScreen(
+    selectedAmenities: List<Amenity>,
+    onToggleAmenity: (Amenity) -> Unit,
     onNext: () -> Unit,
     onBack: () -> Unit,
     onSkip: () -> Unit,
     canNavigateBack: Boolean,
     modifier: Modifier = Modifier
 ) {
+
+    var selectedOptionRadioButton by remember { mutableStateOf("") }
+
     GradientBackground {
+
         Column(
-            modifier = Modifier
+            modifier = modifier
                 .fillMaxSize()
                 .padding(24.dp)
         ) {
-
+            // Progress indicator
             ProgressBar(
-                progress = 1f,
+                progress = 0.8f,
                 modifier = Modifier.fillMaxWidth(),
             )
 
             Spacer(modifier = Modifier.height(40.dp))
 
             Text(
-                text = stringResource(R.string.mainGoalTitle),
+                text = "What amenities are important to you?",
                 fontSize = 20.sp,
                 fontWeight = FontWeight.Bold,
                 color = MaterialTheme.colorScheme.onBackground,
-                fontFamily = Poppins
+                fontFamily = Poppins,
             )
 
             Spacer(modifier = Modifier.height(8.dp))
 
             Text(
-                text = stringResource(R.string.mainGoalDescription),
+                text = "Select all that matter to you (you can choose multiple)",
                 fontSize = 14.sp,
                 color = MaterialTheme.colorScheme.onBackground,
                 fontWeight = FontWeight.Normal,
                 modifier = Modifier.fillMaxWidth()
             )
+
             Spacer(modifier = Modifier.height(32.dp))
 
             LazyColumn(
                 verticalArrangement = Arrangement.spacedBy(16.dp),
                 modifier = Modifier.weight(1f)
             ) {
-                items(MainGoal.entries.toTypedArray()) { goal ->
+                items(Amenity.entries.toTypedArray()) { amenity ->
                     BalkanEstateSelectionCard(
-                        title = goal.displayName,
-                        description = goal.description,
+                        title = amenity.title,
+                        description = amenity.description,
                         isSelected = false,
-                        onClick = {},
+                        onClick = { },
                         selectionType = SelectionType.CHECKBOX,
                         showSelectionIndicator = true
                     )
@@ -94,24 +100,26 @@ fun MainGoalScreen(
 
             Spacer(modifier = Modifier.height(24.dp))
 
+            // Navigation buttons
             Row(
                 modifier = Modifier.fillMaxWidth(),
                 horizontalArrangement = Arrangement.spacedBy(12.dp)
             ) {
-                if(canNavigateBack) {
+                if (canNavigateBack) {
                     BalkanEstateOutlinedActionButton(
-                        text = stringResource(R.string.go_back),
+                        onClick = onBack,
+                        text = "Back",
                         isLoading = false,
                         enabled = true,
-                        onClick = {  },
                         modifier = Modifier.weight(1f)
                     )
                 }
+
                 BalkanEstateActionButton(
-                    text = stringResource(R.string.next),
+                    onClick = onNext,
+                    text = "Next",
                     isLoading = false,
                     enabled = true,
-                    onClick = {  },
                     modifier = Modifier.weight(1f)
                 )
             }
@@ -126,16 +134,19 @@ fun MainGoalScreen(
 
 }
 
+
 @Preview(showBackground = true)
 @Composable
-fun MainGoalScreenPreview() {
+private fun AmenityScreenPreview() {
     BalkanEstateTheme {
-        MainGoalScreen(
-            selectedType = MainGoal.QUICK_SALE,
-            onToggleSelection = { },
+        AmenitiesBuyerScreen(
+            selectedAmenities = listOf(
+                Amenity.GOOD_SCHOOLS, Amenity.PARKS_RECREATION
+            ),
             onNext = { },
-            onBack = { },
+            onBack = {},
             onSkip = { },
+            onToggleAmenity = { },
             canNavigateBack = true,
         )
     }
