+++ onboarding/presentation/src/main/java/com/zanoapps/onboarding/presentation/buyer/currentlifesituation/CurrentLifeSituationActions.kt
@@ -1,0 +1,11 @@
+package com.zanoapps.onboarding.presentation.buyer.currentlifesituation
+
+import com.zanoapps.onboarding.domain.enums.buyer.LifeSituation
+import com.zanoapps.onboarding.domain.enums.buyer.PropertyIntent
+
+sealed interface CurrentLifeSituationActions {
+    data class OnPreferenceSelected(val preference: LifeSituation) : CurrentLifeSituationActions
+    data object OnBackClick : CurrentLifeSituationActions
+    data object OnNextClick : CurrentLifeSituationActions
+    data object OnSkipClick : CurrentLifeSituationActions
+}
