+++ onboarding/presentation/src/main/java/com/zanoapps/onboarding/presentation/buyer/thankyoubuyer/ThankYouAction.kt
@@ -1,0 +1,8 @@
+package com.zanoapps.onboarding.presentation.buyer.thankyoubuyer
+
+
+interface ThankYouAction {
+    // TODO("Object with all details in it")
+    data object OnSearchPropertiesClicked : ThankYouAction
+
+}
