+++ onboarding/presentation/src/main/java/com/zanoapps/onboarding/presentation/seller/pricing/PricingActions.kt
@@ -1,0 +1,7 @@
+package com.zanoapps.onboarding.presentation.seller.pricing
+
+sealed interface PricingActions {
+    data class OnSelectPlanClick(val option: PricingCTAStyle): PricingActions
+    data object OnSkipClicked: PricingActions
+
+}
