+++ onboarding/presentation/src/main/java/com/zanoapps/onboarding/presentation/seller/sellingtime/SellingTimeActions.kt
@@ -1,0 +1,11 @@
+package com.zanoapps.onboarding.presentation.seller.sellingtime
+
+import com.zanoapps.onboarding.domain.enums.seller.SellingTime
+
+sealed interface SellingTimeActions {
+
+    data class OnPreferenceSelected(val preference: SellingTime) : SellingTimeActions
+    data object OnBackClick : SellingTimeActions
+    data object OnNextClick : SellingTimeActions
+    data object OnSkipClick : SellingTimeActions
+}
